move deleteAllWorkflowReminders to repo
