+++ packages/trpc/server/routers/viewer/workflows/delete.handler.ts
@@ -1,10 +1,11 @@
+import { WorkflowRepository } from "@calcom/lib/server/repository/workflow";
 import { prisma } from "@calcom/prisma";
 import type { TrpcSessionUser } from "@calcom/trpc/server/trpc";
 
 import { TRPCError } from "@trpc/server";
 
 import type { TDeleteInputSchema } from "./delete.schema";
-import { deleteAllWorkflowReminders, isAuthorized, removeSmsReminderFieldForEventTypes } from "./util";
+import { isAuthorized, removeSmsReminderFieldForEventTypes } from "./util";
 
 type DeleteOptions = {
   ctx: {
@@ -50,7 +51,7 @@ export const deleteHandler = async ({ ctx, input }: DeleteOptions) => {
   });
 
   //cancel workflow reminders of deleted workflow
-  await deleteAllWorkflowReminders(scheduledReminders);
+  await WorkflowRepository.deleteAllWorkflowReminders(scheduledReminders);
 
   const isOrg = workflowToDelete.team?.isOrganization ?? false;
